Make "struct regmatch_t" match "Tcl_RegExpIndices" (again). Win64 build was broken because of this. Other platforms were OK.

+++ generic/regexec.c
@@ -889,7 +889,7 @@ cbrdissect(
     MDEBUG(("cbackref n%d %d{%d-%d}\n", t->id, n, min, max));
 
     /* get the backreferenced string */
-    if (v->pmatch[n].rm_so == (size_t)-1) {
+    if (v->pmatch[n].rm_so == -1) {
 	return REG_NOMATCH;
     }
     brstring = v->start + v->pmatch[n].rm_so;
